refactor; moved over CDF loaders
