+++ cognite/neat/utils/cdf_loaders.py
